+++ tis-client/src/main/java/com/cas/sim/tis/view/control/imp/library/LibraryList.java
@@ -177,7 +177,7 @@ public class LibraryList extends HBox implements IContent {
 			table.setItems(null);
 			table.build();
 		} else {
-			pagination.setPageCount((int) pageInfo.getPages());
+			pagination.setPageCount(pageInfo.getPages());
 			JSONArray array = new JSONArray();
 			array.addAll(pageInfo.getList());
 			table.setItems(array);
